moving common functions in one file
